Reorganize and delete unused code
